+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,4 +1,4 @@
-## v2.1.0-alpha.0 (Jul 05, 2022)
+## v2.1.0 (Jul 05, 2022)
 
 -   Replace `Mobify` references/links with proper PWA Kit values. [#619](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/619)
 -   Add support for a custom build directory to `pwa-kit-dev build`. [#628](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/628)